@@ -6,11 +6,12 @@ See LICENSE for details
 """
 
 from avro.schema import Schema as AvroSchema
+from collections.abc import Callable
 from karapace.errors import InvalidVersion, VersionNotFoundException
 from karapace.schema_models import parse_avro_schema_definition, SchemaVersion, TypedSchema, Versioner
 from karapace.schema_type import SchemaType
 from karapace.typing import Version, VersionTag
-from typing import Any, Callable, Optional
+from typing import Any
 
 import operator
 import pytest
@@ -91,7 +92,7 @@ class TestVersioner:
         avro_schema: str,
         avro_schema_parsed: AvroSchema,
     ) -> Callable[[Version, dict[str, Any]], dict[Version, SchemaVersion]]:
-        def schema_versions(version: Version, schema_version_data: Optional[dict[str, Any]] = None):
+        def schema_versions(version: Version, schema_version_data: dict[str, Any] | None = None):
             schema_version_data = schema_version_data or dict()
             base_schema_version_data = dict(
                 subject="test-topic",
