@@ -7,8 +7,7 @@ from __future__ import annotations
 from collections.abc import Generator
 from pathlib import Path
 from tempfile import mkstemp, TemporaryDirectory
-from typing import Final, IO, Literal
-from typing_extensions import TypeAlias
+from typing import Final, IO, Literal, TypeAlias
 
 import contextlib
 import os
