@@ -2,9 +2,10 @@
 Copyright (c) 2023 Aiven Ltd
 See LICENSE for details
 """
+from collections.abc import Callable
 from dataclasses import dataclass
 from enum import Enum, unique
-from typing import Callable, Generic, TypeVar
+from typing import Generic, TypeVar
 
 T = TypeVar("T")
 
