@@ -10,8 +10,6 @@ from dataclasses import dataclass
 from typing import TypeVar
 from typing_extensions import dataclass_transform
 
-import sys
-
 T = TypeVar("T")
 
 
@@ -20,11 +18,8 @@ T = TypeVar("T")
     kw_only_default=True,
 )
 def default_dataclass(cls: type[T]) -> type[T]:
-    if sys.version_info >= (3, 10):
-        return dataclass(
-            frozen=True,
-            slots=True,
-            kw_only=True,
-        )(cls)
-
-    return dataclass(frozen=True)(cls)
+    return dataclass(
+        frozen=True,
+        slots=True,
+        kw_only=True,
+    )(cls)
